+++ test/integration/test_b2_command_line.py
@@ -129,11 +129,11 @@ def test_basic(b2_tool, bucket_name):
 
     b2_tool.should_succeed(['hide-file', bucket_name, 'c'])
 
-    list_of_files = b2_tool.should_succeed_json(['ls', '--json', bucket_name])
+    list_of_files = b2_tool.should_succeed_json(['ls', '--json', bucket_name, '**'])
     should_equal(['a', 'b/1', 'b/2', 'd'], [f['fileName'] for f in list_of_files])
 
     list_of_files = b2_tool.should_succeed_json(
-        ['ls', '--json', '--versions', bucket_name]
+        ['ls', '--json', '--versions', bucket_name, '**']
     )
     should_equal(['a', 'a', 'b/1', 'b/2', 'c', 'c', 'd'], [f['fileName'] for f in list_of_files])
     should_equal(
@@ -1141,7 +1141,7 @@ def test_sse_b2(b2_tool, bucket_name):
             ]
         )
 
-    list_of_files = b2_tool.should_succeed_json(['ls', '--json', bucket_name])
+    list_of_files = b2_tool.should_succeed_json(['ls', '--json', bucket_name, '**'])
     should_equal(
         [{
             'algorithm': 'AES256',
@@ -1168,7 +1168,7 @@ def test_sse_b2(b2_tool, bucket_name):
         ['copy-file-by-id', not_encrypted_version['fileId'], bucket_name, 'copied_not_encrypted']
     )
 
-    list_of_files = b2_tool.should_succeed_json(['ls', '--json', bucket_name])
+    list_of_files = b2_tool.should_succeed_json(['ls', '--json', bucket_name, '**'])
     should_equal(
         [{
             'algorithm': 'AES256',